refactor: namespace 및 코드 수정
