+++ src/Laraboard/tests/Feature/Laraboard/CommentDeleteTest.php
@@ -1,14 +1,13 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\Laraboard;
 
 use App\Models\Laraboard\Board;
 use App\Models\Laraboard\Comment;
 use App\Models\Laraboard\Post;
 use App\Models\User;
-use Illuminate\Foundation\Testing\RefreshDatabase;
-use Illuminate\Foundation\Testing\WithFaker;
 use Illuminate\Http\Response;
+use Inium\Laraboard\Support\Traits\Tests\RecursiveRefreshDatabaseTrait as RefreshDatabase;
 use Tests\TestCase;
 
 class CommentDeleteTest extends TestCase
