+++ src/Laraboard/tests/Feature/Laraboard/CommentStoreTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\Laraboard;
 
 use App\Models\Laraboard\Comment;
 use App\Models\Laraboard\Post;
 use App\Models\User;
-use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Foundation\Testing\WithFaker;
+use Inium\Laraboard\Support\Traits\Tests\RecursiveRefreshDatabaseTrait as RefreshDatabase;
 use Illuminate\Http\Response;
 use Tests\TestCase;
