+++ src/Laraboard/tests/Feature/Laraboard/PostNoticeUpdateTest.php
@@ -1,32 +1,36 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\Laraboard;
 
 use App\Models\Laraboard\Board;
 use App\Models\Laraboard\Post;
 use App\Models\User;
-use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Foundation\Testing\WithFaker;
+use Inium\Laraboard\Support\Traits\Tests\RecursiveRefreshDatabaseTrait as RefreshDatabase;
 use Symfony\Component\HttpFoundation\Response;
 use Tests\TestCase;
 
-class PostUpdateTest extends TestCase
+class PostNoticeUpdateTest extends TestCase
 {
     use RefreshDatabase;
     use WithFaker;
 
     /**
-     * 게시글 수정 테스트 - 성공 200 OK
+     * 공지글 수정 테스트
+     * - 성공 200 OK
      *
      * @return void
      */
-    public function test_게시글_수정_성공_200_Ok(): void
+    public function test_공지글_수정_200_Ok(): void
     {
-        $post = Post::factory()->create();
+        $post = Post::factory()
+            ->notice()
+            ->create();
 
         $formData = [
             "subject" => $this->randomHtmlSentences(rand(1, 3)),
             "content" => $this->randomHtmlSentences(rand(20, 50)),
+            "notice" => false,
         ];
 
         $response = $this->actingAs($post->user)->putJson(
@@ -43,7 +47,7 @@ class PostUpdateTest extends TestCase
         $isExist =
             Post::where([
                 ["wrote_user_id", "=", $post->user->id],
-                ["notice", "=", false],
+                ["notice", "=", $formData["notice"]],
                 ["subject", "=", strip_tags($formData["subject"])],
             ])->count() > 0;
 
@@ -51,18 +55,19 @@ class PostUpdateTest extends TestCase
     }
 
     /**
-     * 게시글 수정 테스트 실패
+     * 공지글 수정 테스트 실패
      * - 사용자 정보 없음 401 Unauthorized
      *
      * @return void
      */
-    public function test_게시글_수정_실패_작성자_정보_없음_401_Unauthorized(): void
+    public function test_공지글_수정_실패_작성자_정보_없음_401_Unauthorized(): void
     {
         $post = Post::factory()->create();
 
         $formData = [
             "subject" => $this->randomHtmlSentences(rand(1, 3)),
             "content" => $this->randomHtmlSentences(rand(20, 50)),
+            "notice" => false,
         ];
 
         $response = $this->putJson(
@@ -77,14 +82,16 @@ class PostUpdateTest extends TestCase
     }
 
     /**
-     * 게시글 수정 테스트 실패
+     * 공지글 수정 테스트 실패
      * - 사용자 정보 불치 401 Unauthorized
      *
      * @return void
      */
-    public function test_게시글_수정_실패_게시글_작성자_정보_불일치_401_Unauthorized(): void
+    public function test_공지글_수정_실패_게시글_작성자_정보_불일치_401_Unauthorized(): void
     {
-        $post = Post::factory()->create();
+        $post = Post::factory()
+            ->notice()
+            ->create();
         $user = User::factory()->create();
 
         $formData = [
@@ -104,14 +111,16 @@ class PostUpdateTest extends TestCase
     }
 
     /**
-     * 게시글 수정 테스트 실패
+     * 공지글 수정 테스트 실패
      * - 잘못된 게시판 404 Not Found (form validation fail)
      *
      * @return void
      */
-    public function test_게시글_수정_실패_잘못된_게시판_404_Not_Found(): void
+    public function test_공지글_수정_실패_잘못된_게시판_404_Not_Found(): void
     {
-        $post = Post::factory()->create();
+        $post = Post::factory()
+            ->notice()
+            ->create();
         $board = Board::factory()->create();
 
         $formData = [
@@ -131,14 +140,16 @@ class PostUpdateTest extends TestCase
     }
 
     /**
-     * 게시글 수정 테스트 실패
-     * - 공지여부 잘못된 자료형 값 422 Unprocessable Entity (form validation fail)
+     * 공지글 수정 테스트 실패
+     * - 공지여부 잘못된 자료형 값 422_Unprocessable_Entity (form validation fail)
      *
      * @return void
      */
-    public function test_게시글_수정_실패_공지여부_잘못된값_422_Unprocessable_Entity(): void
+    public function test_공지글_수정_실패_공지여부_잘못된값_422_Unprocessable_Entity(): void
     {
-        $post = Post::factory()->create();
+        $post = Post::factory()
+            ->notice()
+            ->create();
 
         $formData = [
             "subject" => $this->randomHtmlSentences(rand(1, 3)),
@@ -158,18 +169,21 @@ class PostUpdateTest extends TestCase
     }
 
     /**
-     * 게시글 수정 테스트 실패
-     * - 제목 없음 422 Unprocessable Entity (form validation fail)
+     * 공지글 수정 테스트 실패
+     * - 제목 없음 422_Unprocessable_Entity(form validation fail)
      *
      * @return void
      */
-    public function test_게시글_수정_실패_제목없음_422_Unprocessable_Entity(): void
+    public function test_공지글_수정_실패_제목없음_422_Unprocessable_Entity(): void
     {
-        $post = Post::factory()->create();
+        $post = Post::factory()
+            ->notice()
+            ->create();
 
         $formData = [
             // "subject" => $this->randomHtmlSentences(rand(1, 3)),
             "content" => $this->randomHtmlSentences(rand(20, 50)),
+            "notice" => true,
         ];
 
         $response = $this->actingAs($post->user)->putJson(
@@ -184,18 +198,21 @@ class PostUpdateTest extends TestCase
     }
 
     /**
-     * 게시글 수정 테스트 실패
-     * - 본문 없음 422 Unprocessable Entity (form validation fail)
+     * 공지글 수정 테스트 실패
+     * - 본문 없음 422_Unprocessable_Entity (form validation fail)
      *
      * @return void
      */
-    public function test_게시글_수정_실패_본문없음_422_Unprocessable_Entity(): void
+    public function test_공지글_수정_실패_본문없음_422_Unprocessable_Entity(): void
     {
-        $post = Post::factory()->create();
+        $post = Post::factory()
+            ->notice()
+            ->create();
 
         $formData = [
             "subject" => $this->randomHtmlSentences(rand(1, 3)),
             // "content" => $this->randomHtmlSentences(rand(20, 50)),
+            "notice" => true,
         ];
 
         $response = $this->actingAs($post->user)->postJson(
