+++ src/Laraboard/tests/Feature/Laraboard/PostUpdateTest.php
@@ -1,36 +1,32 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\Laraboard;
 
 use App\Models\Laraboard\Board;
 use App\Models\Laraboard\Post;
 use App\Models\User;
-use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Foundation\Testing\WithFaker;
+use Inium\Laraboard\Support\Traits\Tests\RecursiveRefreshDatabaseTrait as RefreshDatabase;
 use Symfony\Component\HttpFoundation\Response;
 use Tests\TestCase;
 
-class PostNoticeUpdateTest extends TestCase
+class PostUpdateTest extends TestCase
 {
     use RefreshDatabase;
     use WithFaker;
 
     /**
-     * 공지글 수정 테스트
-     * - 성공 200 OK
+     * 게시글 수정 테스트 - 성공 200 OK
      *
      * @return void
      */
-    public function test_공지글_수정_200_Ok(): void
+    public function test_게시글_수정_성공_200_Ok(): void
     {
-        $post = Post::factory()
-            ->notice()
-            ->create();
+        $post = Post::factory()->create();
 
         $formData = [
             "subject" => $this->randomHtmlSentences(rand(1, 3)),
             "content" => $this->randomHtmlSentences(rand(20, 50)),
-            "notice" => false,
         ];
 
         $response = $this->actingAs($post->user)->putJson(
@@ -47,7 +43,7 @@ class PostNoticeUpdateTest extends TestCase
         $isExist =
             Post::where([
                 ["wrote_user_id", "=", $post->user->id],
-                ["notice", "=", $formData["notice"]],
+                ["notice", "=", false],
                 ["subject", "=", strip_tags($formData["subject"])],
             ])->count() > 0;
 
@@ -55,19 +51,18 @@ class PostNoticeUpdateTest extends TestCase
     }
 
     /**
-     * 공지글 수정 테스트 실패
+     * 게시글 수정 테스트 실패
      * - 사용자 정보 없음 401 Unauthorized
      *
      * @return void
      */
-    public function test_공지글_수정_실패_작성자_정보_없음_401_Unauthorized(): void
+    public function test_게시글_수정_실패_작성자_정보_없음_401_Unauthorized(): void
     {
         $post = Post::factory()->create();
 
         $formData = [
             "subject" => $this->randomHtmlSentences(rand(1, 3)),
             "content" => $this->randomHtmlSentences(rand(20, 50)),
-            "notice" => false,
         ];
 
         $response = $this->putJson(
@@ -82,16 +77,14 @@ class PostNoticeUpdateTest extends TestCase
     }
 
     /**
-     * 공지글 수정 테스트 실패
+     * 게시글 수정 테스트 실패
      * - 사용자 정보 불치 401 Unauthorized
      *
      * @return void
      */
-    public function test_공지글_수정_실패_게시글_작성자_정보_불일치_401_Unauthorized(): void
+    public function test_게시글_수정_실패_게시글_작성자_정보_불일치_401_Unauthorized(): void
     {
-        $post = Post::factory()
-            ->notice()
-            ->create();
+        $post = Post::factory()->create();
         $user = User::factory()->create();
 
         $formData = [
@@ -111,16 +104,14 @@ class PostNoticeUpdateTest extends TestCase
     }
 
     /**
-     * 공지글 수정 테스트 실패
+     * 게시글 수정 테스트 실패
      * - 잘못된 게시판 404 Not Found (form validation fail)
      *
      * @return void
      */
-    public function test_공지글_수정_실패_잘못된_게시판_404_Not_Found(): void
+    public function test_게시글_수정_실패_잘못된_게시판_404_Not_Found(): void
     {
-        $post = Post::factory()
-            ->notice()
-            ->create();
+        $post = Post::factory()->create();
         $board = Board::factory()->create();
 
         $formData = [
@@ -140,16 +131,14 @@ class PostNoticeUpdateTest extends TestCase
     }
 
     /**
-     * 공지글 수정 테스트 실패
-     * - 공지여부 잘못된 자료형 값 422_Unprocessable_Entity (form validation fail)
+     * 게시글 수정 테스트 실패
+     * - 공지여부 잘못된 자료형 값 422 Unprocessable Entity (form validation fail)
      *
      * @return void
      */
-    public function test_공지글_수정_실패_공지여부_잘못된값_422_Unprocessable_Entity(): void
+    public function test_게시글_수정_실패_공지여부_잘못된값_422_Unprocessable_Entity(): void
     {
-        $post = Post::factory()
-            ->notice()
-            ->create();
+        $post = Post::factory()->create();
 
         $formData = [
             "subject" => $this->randomHtmlSentences(rand(1, 3)),
@@ -169,21 +158,18 @@ class PostNoticeUpdateTest extends TestCase
     }
 
     /**
-     * 공지글 수정 테스트 실패
-     * - 제목 없음 422_Unprocessable_Entity(form validation fail)
+     * 게시글 수정 테스트 실패
+     * - 제목 없음 422 Unprocessable Entity (form validation fail)
      *
      * @return void
      */
-    public function test_공지글_수정_실패_제목없음_422_Unprocessable_Entity(): void
+    public function test_게시글_수정_실패_제목없음_422_Unprocessable_Entity(): void
     {
-        $post = Post::factory()
-            ->notice()
-            ->create();
+        $post = Post::factory()->create();
 
         $formData = [
             // "subject" => $this->randomHtmlSentences(rand(1, 3)),
             "content" => $this->randomHtmlSentences(rand(20, 50)),
-            "notice" => true,
         ];
 
         $response = $this->actingAs($post->user)->putJson(
@@ -198,21 +184,18 @@ class PostNoticeUpdateTest extends TestCase
     }
 
     /**
-     * 공지글 수정 테스트 실패
-     * - 본문 없음 422_Unprocessable_Entity (form validation fail)
+     * 게시글 수정 테스트 실패
+     * - 본문 없음 422 Unprocessable Entity (form validation fail)
      *
      * @return void
      */
-    public function test_공지글_수정_실패_본문없음_422_Unprocessable_Entity(): void
+    public function test_게시글_수정_실패_본문없음_422_Unprocessable_Entity(): void
     {
-        $post = Post::factory()
-            ->notice()
-            ->create();
+        $post = Post::factory()->create();
 
         $formData = [
             "subject" => $this->randomHtmlSentences(rand(1, 3)),
             // "content" => $this->randomHtmlSentences(rand(20, 50)),
-            "notice" => true,
         ];
 
         $response = $this->actingAs($post->user)->postJson(
